Optimize container image build
We split the process into 2 stages, build and runner,
in the `build` process, we install all the requirements
which we need for installing the python packages.

And then we just packages the results into the `runner`
stage, which drop all non required packages for running
the tests them self.

This reduces the image size from ~480MB to ~220MB

+ Multistage docker build
+ Remove .git and other unnecessary files from the image

Closes Jira issue: CNV-24271

Change-Id: I0cc82311927e2f8914dfd3572483907d882f09c0
Reviewed-on: https://code.engineering.redhat.com/gerrit/c/cnv-tests/+/427981
Continuous-Integration: CNV-QE Gerrit <no-reply-cnv-qe-devops@redhat.com>
Reviewed-by: Denys Shchedrivyi <dshchedr@redhat.com>
Reviewed-by: Sarah Bennert <sbennert@redhat.com>
Reviewed-by: Meni Yakove <myakove@redhat.com>
Tested-by: Rabin Yasharzadehe <rabin@redhat.com>